Rename evals/basis_eval.py to evals/basis_deriv.py
Change the code, docstrings, notes and file names to reflect this renaming.
