+++ tests/test_basis_deriv.py
@@ -3,7 +3,7 @@ import itertools as it
 
 from gbasis.contractions import GeneralizedContractionShell
 from gbasis.evals._deriv import _eval_deriv_contractions
-from gbasis.evals.eval_deriv import EvalDeriv, evaluate_deriv_basis
+from gbasis.evals.basis_deriv import EvalDeriv, evaluate_deriv_basis
 from gbasis.parsers import make_contractions, parse_nwchem
 import numpy as np
 import pytest
@@ -12,7 +12,7 @@ from utils import find_datafile
 
 
 def test_evaluate_deriv_construct_array_contraction():
-    """Test gbasis.evals.evaluate_deriv.EvalDeriv.construct_array_contraction."""
+    """Test gbasis.evals.basis_deriv.EvalDeriv.construct_array_contraction."""
     points = np.array([[2, 3, 4]])
     orders = np.array([0, 0, 0])
     contractions = GeneralizedContractionShell(
@@ -129,7 +129,7 @@ def test_evaluate_deriv_construct_array_contraction():
 
 
 def test_evaluate_deriv_basis_cartesian():
-    """Test gbasis.evals.basis_eval.evaluate_deriv_basis_cartesian."""
+    """Test gbasis.evals.basis_deriv.evaluate_deriv_basis_cartesian."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
     evaluate_obj = EvalDeriv(basis)
@@ -152,7 +152,7 @@ def test_evaluate_deriv_basis_cartesian():
 
 
 def test_evaluate_deriv_basis_spherical():
-    """Test gbasis.evals.basis_eval.evaluate_deriv_basis_spherical."""
+    """Test gbasis.evals.basis_deriv.evaluate_deriv_basis_spherical."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
     evaluate_obj = EvalDeriv(basis)
@@ -175,7 +175,7 @@ def test_evaluate_deriv_basis_spherical():
 
 
 def test_evaluate_deriv_basis_mix():
-    """Test gbasis.evals.basis_eval.evaluate_deriv_basis_mix."""
+    """Test gbasis.evals.basis_deriv.evaluate_deriv_basis_mix."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
     evaluate_obj = EvalDeriv(basis)
@@ -198,7 +198,7 @@ def test_evaluate_deriv_basis_mix():
 
 
 def test_evaluate_deriv_basis_lincomb():
-    """Test gbasis.evals.basis_eval.evaluate_deriv_basis_lincomb."""
+    """Test gbasis.evals.basis_deriv.evaluate_deriv_basis_lincomb."""
     basis_dict = parse_nwchem(find_datafile("data_sto6g.nwchem"))
     basis = make_contractions(basis_dict, ["Kr"], np.array([[0, 0, 0]]))
     evaluate_obj = EvalDeriv(basis)
